requirements.txt
Django
pubnub
djangorestframework
django-filter